requirements.txt
absl-py==0.8.1
appnope==0.1.0
asgiref==3.2.3
asn1crypto==0.24.0
astor==0.8.0
attrs==19.2.0
Automat==0.7.0
backcall==0.1.0
bleach==3.1.0
cachetools==3.1.1
certifi==2017.7.27.1
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.2.2
decorator==4.4.0
defusedxml==0.6.0
Django==3.0.1
django-cors-headers==3.2.0
django-node==4.0.0
django-rest-framework==0.1.0
django-rest-knox==4.1.0
django-webpack-loader==0.6.0
djangorestframework==3.11.0
entrypoints==0.3
gast==0.2.2
google-auth==1.7.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.7
grpcio==1.25.0
h5py==2.10.0
hyperlink==19.0.0
idna==2.6
incremental==17.5.0
ipdb==0.12.2
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.3
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.5.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
numpy==1.17.2
oauthlib==3.1.0
opt-einsum==3.1.0
optional-django==0.1.0
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.10.0
ptyprocess==0.6.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.18
Pygments==2.4.2
PyHamcrest==1.9.0
pyOpenSSL==17.5.0
pyrsistent==0.15.4
python-dateutil==2.8.0
pytz==2017.2
pyzmq==18.1.0
qtconsole==4.5.5
requests==2.18.4
requests-oauthlib==1.3.0
rsa==4.0
Send2Trash==1.5.0
sentencepiece==0.1.85
six==1.11.0
sqlparse==0.3.0
tensorboard==2.0.1
tensorflow==2.0.0
tensorflow-estimator==2.0.1
tensorflow-hub==0.6.0
tensorflow-serving-api==2.0.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.3
Twisted==19.7.0
urllib3==1.22
v==0.0.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zope.interface==4.6.0
